Add/Fix - minor changes on scenes
Both Addestramento and Benvenuto scenes have now a Manager GameObject, this is meant to be used for purely logic purposes.

Benvenuto has an AudioSource component in Camera GameObject.

Addestramento is *unloaded* by default, resulting in fixing the warning *"There are 2 audio listeners in the scene. Please ensure there is always exactly one audio listener in the scene."*